pyproject.toml
[project]
name = "fasthtml-fungi-2"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Daniel N. Lang", email = "daniel.lang@velaia.de" }
]
dependencies = [
    "python-fasthtml>=0.3.3",
    "icecream>=2.1.3",
    "piexif>=1.1.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"fasthtml-fungi-2" = "fasthtml_fungi_2:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fasthtml_fungi_2"]
